Rename dissertation related pages to kebab-case for consistency
