+++ frontend/src/pages/dissertation-detail-page.tsx
