+++ frontend/src/pages/dissertation-list-page.tsx
